Add passthrough of block device mapping from image into instance
Why doesn't AWS do this already? Don't know, see
http://docs.aws.amazon.com/AWSEC2/latest/UserGuide/block-device-mapping-concepts.html
"For M3 instances, you must specify instance store volumes in the block
device mapping for the instance when you launch it. When you launch an
M3 instance, instance store volumes specified in the block device
mapping for the AMI may be ignored if they are not specified as part of
the instance block device mapping."